Op.34 No.1: add missing \alternative block for rightHandLower
I have not done the research to determine why LilyPond behaves in this
way, but when rightHandLower lacked an \alternative within "\repeat
volta 2 { ... }", an additional repeat barline was added to the score in
the 2nd ending.

The solution is to add the missing \alternative block for
rightHandLower, and the barline between measures 49 (volta 2 measure)
and 50 becomes a normal one.